+++ Frontend/app/crm.module.ts
@@ -6,38 +6,41 @@ import {HttpModule} from '@angular/http';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {RouterModule, Routes} from '@angular/router';
 import {ToastModule} from 'ng2-toastr';
-import {MainComponent} from './main/main.component';
-import {MobileNavigationComponent} from './main/ui/mobile/m-navigation.component';
+import {MainComponent} from './main/components/main.component';
 import {MobileDashboardComponent} from './main/ui/mobile-dashboard.component';
 import {NotFoundComponent} from './main/routes/not-found.component';
-import {SharedModule} from './shared/shared.module';
-import {UsersModule} from './users/users.module';
-import {StoreModule} from './store/store.module';
 import {DynamicFormsModule} from './forms/dynamic-forms.module';
-import {SideMenuComponent} from './main/ui/mobile/side-menu.component';
+import {SideMenuComponent} from './main/components/side-menu.component';
 import {CRMService} from './main/services/crm.service';
 import {ContactComponent} from './main/components/contacts.component';
 import {QuotesComponent} from './main/components/quotes.component';
 import {NotesComponent} from './main/components/notes.component';
-import {UIService} from './main/services/ui.service';
 import {BottomMenuComponent} from './main/ui/bottom-menu.component';
-import {SmartFormComponent} from './forms/components/smart-form.component';
 import * as _ from 'lodash';
+import {HomeComponent} from './main/components/home.component';
+import {CompaniesComponent} from './main/components/companies.component';
+import {AddCompanyComponent} from './main/components/add-company.component';
+import {AddContactComponent} from './main/components/add-contact.component';
 
 const MAIN_ROUTES: Routes = [
-	{path:'main', component: MainComponent},
-	{path:'create', component: SmartFormComponent},
-	{path:'details', component: SmartFormComponent},
-	{path:'Contacts', component: ContactComponent},
-	{path:'Quotes', component: QuotesComponent},
-	{path:'', redirectTo: '/Contacts', pathMatch: 'full'},
+	{path: 'main', component: MainComponent},
+	{path: 'Home', component: HomeComponent},
+	{path: 'Companies', component: CompaniesComponent},
+	{path: 'Add-Company', component: AddCompanyComponent},
+	{path: 'Add-Contact/:owner_id', component: AddContactComponent},
+	{path: 'Contacts', component: ContactComponent},
+	{path: 'Quotes', component: QuotesComponent},
+	{path: '', redirectTo: '/Home', pathMatch: 'full'},
 	{path: '**', component: NotFoundComponent}
 ];
 
 const COMPONENTS = [
+	HomeComponent,
+	AddContactComponent,
+	AddCompanyComponent,
+	CompaniesComponent,
 	SideMenuComponent,
 	MainComponent,
-	MobileNavigationComponent,
 	MobileDashboardComponent,
 	NotFoundComponent,
 	ContactComponent,
@@ -54,12 +57,9 @@ const MODULES = [
 	BrowserModule,
 	HttpModule,
 	ToastModule.forRoot(),
-	UsersModule,
-	SharedModule,
 	FormsModule,
 	ReactiveFormsModule,
 	DynamicFormsModule,
-	StoreModule,
 ];
 
 @NgModule({
@@ -71,7 +71,6 @@ const MODULES = [
 	],
 	providers: [
 		CRMService,
-		UIService,
 	],
 	schemas: [
 		CUSTOM_ELEMENTS_SCHEMA