+++ Frontend/app/forms/dynamic-forms.module.ts
@@ -5,21 +5,17 @@ import {InputComponent} from './components/input.component';
 import {FormComponent} from './components/form.component';
 import {CommonModule} from '@angular/common';
 import {RouterModule, Routes} from '@angular/router';
-import {SmartFormComponent} from './components/smart-form.component';
 import {FormsService} from './services/forms.service';
 import {HAMMER_GESTURE_CONFIG, HammerGestureConfig} from '@angular/platform-browser';
 import {SharedModule} from '../shared/shared.module';
 
 const ROUTES: Routes = [
-	{path: 'create', component: SmartFormComponent},
-	{path: 'details', component: SmartFormComponent}
 ];
 
 const COMPONENTS = [
 	ListComponent,
 	InputComponent,
 	FormComponent,
-	SmartFormComponent,
 ];
 
 const MODULES = [