+++ Frontend/app/main/components/home.component.ts
@@ -1,0 +1,12 @@
+import {Component} from '@angular/core';
+@Component({
+	selector: 'home-component',
+	template:`
+	<h1>Home Componnent</h1>
+		Welcome to Angular Bros. CRM
+	`
+})
+
+export class HomeComponent {
+
+}