+++ Frontend/app/main/models/company.model.ts
@@ -1,5 +1,6 @@
 import {_baseListItem} from './base-model.model';
 import {Contact} from './contact.model';
+import {Quote} from './quote.model';
 
 export interface Company extends _baseListItem{
 	id?: string
@@ -13,26 +14,27 @@ export interface Company extends _baseListItem{
 	misc: string;
 	fax: string;
 	contacts: Contact[];
+	quotes: Quote[];
 	modelType: 'companies';
 	singular: 'Company';
 
 }
 
-
 export function newCompany(): Company {
-		return {
-			id: null,
-			name: null,
-			addressOne: null,
-			addressTwo: null,
-			city: null,
-			zip: null,
-			phone: null,
-			web: null,
-			misc: null,
-			fax: null,
-			contacts: [],
-			modelType: 'companies',
-			singular: 'Company'
-		}
+	return {
+		id: null,
+		name: null,
+		addressOne: null,
+		addressTwo: null,
+		city: null,
+		zip: null,
+		phone: null,
+		web: null,
+		misc: null,
+		fax: null,
+		contacts: [],
+		quotes: [],
+		modelType: 'companies',
+		singular: 'Company'
+	}
 }