+++ Frontend/app/store/models/typescript-cache.model.ts
@@ -1,5 +1,0 @@
-import {List} from '../../forms/models/lists.model';
-
-export interface RDCache {
-	[name: string]: any
-}